스크립트 라임라인 double 형을 반환

+++ src/main/java/TubeSlice/tubeSlice/domain/text/Text.java
@@ -20,7 +20,7 @@ public class Text {
     @Column(name = "id")
     private Long id;
 
-    private Integer timeline;
+    private Double timeline;
 
     private String scripts;
